Update MedicalQueryParser.py
Fixed a small edit I had made local... self._maxHits recognizes that we're not using a document collector and that the total hits are unknown. Originally the algo read 1, the problem with this is the lower you get in the cascade the more similar concept text looks and esp when you hit step 5 with the alpha sorted lancaster stemmer many results return, not just one. Also fixed a small bug where query2 should have read query3.